Move ComponentToFloat and FloatToComponent
Instead of duplicating the definitions of decoding functions as
DecodeToFloat, move ComponentToFloat to ResourceReaderUtils.

This change also moves FloatToComponent for symmetry.
