+++ GLTFSDK/Source/GLTFResourceReader.cpp
@@ -2,17 +2,12 @@
 // Licensed under the MIT License.
 
 #include <GLTFSDK/GLTFResourceReader.h>
+#include <GLTFSDK/ResourceReaderUtils.h>
 
 using namespace Microsoft::glTF;
 
 namespace
 {
-    // Conversions of normalized component types to/from floats are explicitly defined in the 2.0 spec
-    float DecodeToFloat(const int8_t w)  { return std::max(static_cast<float>(w) / 127.0f, -1.0f); }
-    float DecodeToFloat(const uint8_t w) { return static_cast<float>(w) / 255.0f; }
-    float DecodeToFloat(const int16_t w) { return std::max(static_cast<float>(w) / 32767.0f, -1.0f); }
-    float DecodeToFloat(const uint16_t w){ return static_cast<float>(w) / 65535.0f; }
-
     template<typename T>
     std::vector<float> DecodeToFloats(const Document& doc, const GLTFResourceReader& reader, const Accessor& accessor)
     {
@@ -24,7 +19,7 @@ namespace
         if (accessor.normalized)
         {
             std::transform(rawData.begin(), rawData.end(), std::back_inserter(floatData),
-                [](T value) -> float { return DecodeToFloat(value); });
+                [](T value) -> float { return ComponentToFloat(value); });
         }
         else
         {
